+++ src/ac_optimize_ag_reactivity.cpp
@@ -2,6 +2,8 @@
 #include <RcppArmadillo.h>
 #include "acmap_optimization.h"
 #include "acmap_titers.h"
+#include "ac_optimization.h"
+#include "ac_optim_map_stress.h"
 
 // [[Rcpp::export]]
 double ac_reactivity_adjustment_stress(
@@ -16,7 +18,10 @@ double ac_reactivity_adjustment_stress(
     const arma::uvec &fixed_antigens,
     const arma::uvec &fixed_sera,
     const arma::mat &titer_weights,
-    const double &reactivity_stress_weighting
+    const double &reactivity_stress_weighting,
+    const bool reoptimize,
+    const arma::uword num_optimizations,
+    const double &dilution_stepsize
 ) {
 
   // Update reactivities with values from par
@@ -24,20 +29,48 @@ double ac_reactivity_adjustment_stress(
   ag_reactivity_adjustments.elem(arma::find_nonfinite(ag_reactivity_adjustments)) = par;
 
   // Get adjusted map stress
-  double stress = ac_relax_coords(
-    titertable.numeric_table_distances(
+  double stress;
+  if (reoptimize) {
+
+    // Do not report progress
+    options.report_progress = false;
+
+    // Run the optimization
+    std::vector<AcOptimization> optimizations;
+    optimizations = ac_runOptimizations(
+      titertable,
       minimum_column_basis,
       fixed_column_bases,
-      ag_reactivity_adjustments
-    ),
-    titertable.get_titer_types(),
-    ag_coords,
-    sr_coords,
-    options,
-    fixed_antigens,
-    fixed_sera,
-    titer_weights
-  );
+      ag_reactivity_adjustments,
+      ag_coords.n_cols,
+      num_optimizations,
+      options,
+      titer_weights,
+      dilution_stepsize
+    );
+
+    // Sort by stress and keep lowest stress
+    sort_optimizations_by_stress(optimizations);
+    stress = optimizations[0].stress;
+
+  } else {
+
+    stress = ac_relax_coords(
+      titertable.numeric_table_distances(
+        minimum_column_basis,
+        fixed_column_bases,
+        ag_reactivity_adjustments
+      ),
+      titertable.get_titer_types(),
+      ag_coords,
+      sr_coords,
+      options,
+      fixed_antigens,
+      fixed_sera,
+      titer_weights
+    );
+
+  }
 
   // Increase map stress according to adjustment applied
   for (arma::uword i=0; i<ag_reactivity_adjustments.n_elem; i++) {