+++ NEWS.md
@@ -63,3 +63,4 @@
 * Add option to optimize from scratch when performing optimizeAgReactivity
 * Fix an error when performing procrustes with some coordinates set to NA
 * Fix an error where transforming na coordinates in the viewer would give a numeric value
+* Add an option to use arrowheads rather than arrows to show outlier points when plotting